Fix float64 2 uint

+++ handlers/questions.go
@@ -161,23 +161,23 @@ func GetQuestionList(c *gin.Context) {
 		}
 
 		// Create a map for easy lookup
-		scoreMap := make(map[float64]float64)
+		scoreMap := make(map[uint]float64)
 
 		// 填入資料
 		for _, score := range topScores {
 			if score.TopScore != nil {
 				// *int → float64
-				scoreMap[float64(score.QuestionID)] = float64(*score.TopScore)
+				scoreMap[score.QuestionID] = float64(*score.TopScore)
 			} else {
 				// 沒分數給 0.0（或乾脆略過）
-				scoreMap[float64(score.QuestionID)] = 0.0
+				scoreMap[score.QuestionID] = 0.0
 			}
 		}
 
 		// Convert to response format and add user-specific data
 		var responseQuestions []_GetQuestionListQuestionData
 		for _, q := range questions {
-			v := scoreMap[float64(q.ID)]
+			v := scoreMap[q.ID]
 			responseQ := _GetQuestionListQuestionData{
 				Question:    q,
 				HasQuestion: userQuestionMap[q.ID],
